requirements.txt
openai
typer[all]
python-dotenv
requests 